feat: extract out unit card used in curriculum options
